move open console and browser to their separate controls
