+++ app/config.js
@@ -1,4 +1,5 @@
 var path = require('path')
+var os = require('os')
 
 exports = module.exports = {
   'menu-bar-width': 240,
@@ -7,5 +8,9 @@ exports = module.exports = {
     'auto-hide-menu-bar': true,
     width: 800,
     height: 600
-  }
+  },
+  'tray-icon': (os.platform() !== 'darwin' ?
+    path.resolve(__dirname, '../node_modules/ipfs-logo/ipfs-logo-256-ice.png')
+    : path.resolve(__dirname, '../node_modules/ipfs-logo/platform-icons/osx-menu-bar.png')),
+  'webui-path': '/webui'
 }
